requirements.txt
# Python >= 3.7 required
torch>=1.7.0
numpy>=1.19.2
matplotlib>=3.3.2
tqdm>=4.50.2
pillow>=8.0.0  # For GIF creation in visualization